pyproject.toml
[project]
name = "travel-photo-workflow"
version = "1.0.0"
description = "AI-Powered Agentic System for Intelligent Travel Photo Management"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Travel Photo Workflow Team" }
]
keywords = ["ai", "photo-organization", "computer-vision", "travel", "image-processing", "agentic-workflow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core dependencies
    "pillow>=10.1.0",
    "piexif>=1.1.3",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pillow-heif>=0.7.0",
    # Image processing and analysis
    "imagehash>=4.3.1",
    "scikit-image>=0.22.0",
    "scipy>=1.11.0",
    # Machine learning models
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.35.0",
    "sentence-transformers>=2.2.2",
    "open-clip-torch>=2.23.0",
    # API clients for VLM services
    "openai>=1.6.0",
    "google-genai>=0.1.0",
    # Data handling
    "pandas>=2.1.0",
    "python-dateutil>=2.8.2",
    # Utilities
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "jsonschema>=4.20.0",
    # Geolocation
    "geopy>=2.4.0",
    "reverse-geocoder>=1.5.1",
    # Web framework (optional, for serving)
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "dotenv>=0.9.9",
    "python-multipart>=0.0.20",
    "fastapi>=0.127.0",
    "uvicorn>=0.39.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
travel-workflow = "orchestrator:main"

[project.urls]
Homepage = "https://github.com/ycs-mb/Travel-website"
Documentation = "https://github.com/ycs-mb/Travel-website/blob/main/WORKFLOW_DESIGN.md"
Repository = "https://github.com/ycs-mb/Travel-website"
Issues = "https://github.com/ycs-mb/Travel-website/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "utils"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.uv.sources]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | output
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pypackages__",
    "build",
    "dist",
    "output",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=agents --cov=utils --cov=orchestrator --cov-report=term-missing"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
exclude = [
    "output/",
    "build/",
    "dist/",
]
